wip: get atom state value
